backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
pymongo==4.15.2
certifi==2025.10.5
twilio==9.8.4
gunicorn==23.0.0
Pillow==12.1.1
dj-database-url==3.1.2
psycopg2-binary==2.9.11
whitenoise==6.12.0
